backend/pyproject.toml
[tool.poetry]
name = "djangoresful-vue-full-stack"
version = "0.1.0"
description = ""
authors = ["hanson190505 <413506012@qq.com>"]

#poetry install 使用国内源头
[[tool.poetry.source]]
name = "douban"
default = true
url = "https://pypi.doubanio.com/simple/"

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.1.7"
djangorestframework = "^3.12.2"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
